Basic start
I added the packages for front and backend.

I built a simple App.js and css for the front end and connected it to the backends.

I created an index.js for connecting to the routes/stocks.js which is a class for stock data which is added to a watchlist. Connected to an api.